+++ src/SpaceGluer.php
@@ -1,0 +1,8 @@
+<?php declare(strict_types=1);
+
+namespace Jawira\CaseConverter;
+
+abstract class SpaceGluer extends Gluer
+{
+    const DELIMITER = ' ';
+}