+++ src/UppercaseGluer.php
@@ -1,0 +1,8 @@
+<?php declare(strict_types=1);
+
+namespace Jawira\CaseConverter;
+
+abstract class UppercaseGluer extends Gluer
+{
+    const DELIMITER = '';
+}