+++ tests/phpunit/ConvertTest.php
@@ -1,6 +1,23 @@
 <?php
 
+use Jawira\CaseConverter\AdaCase;
+use Jawira\CaseConverter\CamelCase;
+use Jawira\CaseConverter\CobolCase;
 use Jawira\CaseConverter\Convert;
+use Jawira\CaseConverter\DashSplitter;
+use Jawira\CaseConverter\Gluer;
+use Jawira\CaseConverter\KebabCase;
+use Jawira\CaseConverter\LowerCase;
+use Jawira\CaseConverter\MacroCase;
+use Jawira\CaseConverter\PascalCase;
+use Jawira\CaseConverter\SentenceCase;
+use Jawira\CaseConverter\SnakeCase;
+use Jawira\CaseConverter\SpaceSplitter;
+use Jawira\CaseConverter\TitleCase;
+use Jawira\CaseConverter\TrainCase;
+use Jawira\CaseConverter\UnderscoreSplitter;
+use Jawira\CaseConverter\UpperCase;
+use Jawira\CaseConverter\UppercaseSplitter;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -23,12 +40,12 @@ class ConvertTest extends TestCase
         // Disable constructor mocking one method
         $mock = $this->getMockBuilder(Convert::class)
                      ->disableOriginalConstructor()
-                     ->setMethods(['detectNamingConvention'])
+                     ->setMethods(['extractWords'])
                      ->getMock();
 
         // Configuring stub
         $mock->expects($this->once())
-             ->method('detectNamingConvention')
+             ->method('extractWords')
              ->with($this->equalTo('hello_world'));
 
         $class = new ReflectionObject($mock);
@@ -84,7 +101,9 @@ class ConvertTest extends TestCase
      *
      * \Jawira\CaseConverter\Convert::analyse should return Convert::SNAKE if $input contains '_'.
      *
-     * @covers       \Jawira\CaseConverter\Convert::analyse()
+     * @covers       \Jawira\CaseConverter\Convert::analyse
+     * @covers       \Jawira\CaseConverter\Splitter::__construct
+     *
      * @dataProvider analyseProvider
      *
      * @param bool   $isUppercaseWordReturn Times that `isUppercaseWord()` method is called.
@@ -95,7 +114,7 @@ class ConvertTest extends TestCase
      */
     public function testAnalyse(bool $isUppercaseWordReturn, string $expected, string $input)
     {
-        // Disabling constructor without stub methods
+        // Disabling constructor with one stub method
         $stub = $this->getMockBuilder(Convert::class)
                      ->disableOriginalConstructor()
                      ->setMethods(['isUppercaseWord'])
@@ -113,150 +132,25 @@ class ConvertTest extends TestCase
 
         // Testing
         $output = $method->invoke($stub, $input);
-        $this->assertSame($expected, $output);
+        $this->assertInstanceOf($expected, $output);
     }
 
     public function analyseProvider()
     {
         return [
-            'Underscore 1' => [false, Convert::STRATEGY_UNDERSCORE, 'hola_mundo'],
-            'Underscore 2' => [false, Convert::STRATEGY_UNDERSCORE, 'HELLO_WORLD'],
-            'Underscore 3' => [true, Convert::STRATEGY_UNDERSCORE, 'Ñ'],
-            'Underscore 4' => [true, Convert::STRATEGY_UNDERSCORE, 'HELLO'],
-            'Uppercase 1'  => [false, Convert::STRATEGY_UPPERCASE, ''],
-            'Uppercase 2'  => [false, Convert::STRATEGY_UPPERCASE, 'ñ'],
-            'Uppercase 3'  => [false, Convert::STRATEGY_UPPERCASE, 'one'],
-            'Uppercase 4'  => [false, Convert::STRATEGY_UPPERCASE, 'helloWorld'],
-            'Dash 1'       => [false, Convert::STRATEGY_DASH, 'hello-World'],
-            'Dash 2'       => [false, Convert::STRATEGY_DASH, 'my-name-is-bond'],
-            'Space 1'      => [false, Convert::STRATEGY_SPACE, 'Hola mundo'],
-            'Space 2'      => [false, Convert::STRATEGY_SPACE, 'Mi nombre es bond'],
-            'Space 3'      => [false, Convert::STRATEGY_SPACE, 'Formule courte spéciale été'],
-        ];
-    }
-
-    /**
-     * @covers       \Jawira\CaseConverter\Convert::splitString()
-     * @dataProvider splitStringProvider
-     *
-     * @param string $pattern
-     * @param string $input
-     * @param array  $expected
-     *
-     * @throws \ReflectionException
-     */
-    public function testSplitString(string $pattern, string $input, array $expected)
-    {
-        // Disabling constructor without stub methods
-        $mock = $this->getMockBuilder(Convert::class)
-                     ->disableOriginalConstructor()
-                     ->setMethods()
-                     ->getMock();
-
-        // Making public a protected method
-        $reflection = new ReflectionObject($mock);
-        $method     = $reflection->getMethod('splitString');
-        $method->setAccessible(true);
-
-        // Testing method
-        $output = $method->invoke($mock, $pattern, $input);
-        $this->assertSame($expected, $output);
-    }
-
-    public function splitStringProvider()
-    {
-        return [
-            [Convert::DASH, 'hello-world', ['hello', 'world']],
-            [Convert::DASH, 'HeLlO-WoRlD', ['HeLlO', 'WoRlD']],
-            [Convert::DASH, 'Hello-World', ['Hello', 'World']],
-            [Convert::DASH, 'HELLO-WORLD', ['HELLO', 'WORLD']],
-            [Convert::DASH, '--hello--world--', ['hello', 'world']],
-            [Convert::UNDERSCORE, 'hello_world', ['hello', 'world']],
-            [Convert::UNDERSCORE, 'HeLlO_WoRlD', ['HeLlO', 'WoRlD']],
-            [Convert::UNDERSCORE, 'Hello_World', ['Hello', 'World']],
-            [Convert::UNDERSCORE, 'HELLO_WORLD', ['HELLO', 'WORLD']],
-            [Convert::UNDERSCORE, '__hello_____world__', ['hello', 'world']],
-            [Convert::SPACE, 'hEllO wOrlD', ['hEllO', 'wOrlD']],
-            [Convert::SPACE, 'hEllO wOrlD', ['hEllO', 'wOrlD']],
-            [Convert::SPACE, 'hEllO      wOrlD', ['hEllO', 'wOrlD']],
-            [Convert::SPACE, '           hEllO      wOrlD', ['hEllO', 'wOrlD']],
-            [Convert::SPACE, '           hEllO      wOrlD   ', ['hEllO', 'wOrlD']],
-        ];
-    }
-
-    /**
-     * @covers       \Jawira\CaseConverter\Convert::glueString()
-     * @dataProvider glueStringProvider
-     *
-     * @param array  $words
-     * @param string $glue
-     * @param int    $mode
-     * @param bool   $lcf
-     *
-     * @param string $expected
-     *
-     * @throws \ReflectionException
-     */
-    public function testGlueString(array $words, string $glue, int $mode, bool $lcf, string $expected)
-    {
-        // Disabling constructor without stub methods
-        $mock = $this->getMockBuilder(Convert::class)
-                     ->disableOriginalConstructor()
-                     ->setMethods()
-                     ->getMock();
-
-        // Making "words" property accessible and setting a value
-        $reflection = new ReflectionObject($mock);
-        $property   = $reflection->getProperty('words');
-        $property->setAccessible(true);
-        $property->setValue($mock, $words);
-
-        // Making public a protected method
-        $reflection = new ReflectionObject($mock);
-        $method     = $reflection->getMethod('glueString');
-        $method->setAccessible(true);
-
-        // Testing
-        $output = $method->invoke($mock, $glue, $mode, $lcf);
-        $this->assertSame($expected, $output);
-    }
-
-    /**
-     * @return array
-     */
-    public function glueStringProvider()
-    {
-        return [
-            [['foo', 'bar'], Convert::DASH, MB_CASE_LOWER, false, 'foo-bar'],
-            [['foo', 'bar'], Convert::DASH, MB_CASE_TITLE, false, 'Foo-Bar'],
-            [['foo', 'bar'], Convert::DASH, MB_CASE_UPPER, false, 'FOO-BAR'],
-            [['foo', 'bar'], Convert::UNDERSCORE, MB_CASE_LOWER, false, 'foo_bar'],
-            [['foo', 'bar'], Convert::UNDERSCORE, MB_CASE_TITLE, false, 'Foo_Bar'],
-            [['foo', 'bar'], Convert::UNDERSCORE, MB_CASE_UPPER, false, 'FOO_BAR'],
-            [['foo', 'bar'], Convert::EMPTY_STRING, MB_CASE_LOWER, false, 'foobar'],
-            [['foo', 'bar'], Convert::EMPTY_STRING, MB_CASE_TITLE, false, 'FooBar'],
-            [['foo', 'bar'], Convert::EMPTY_STRING, MB_CASE_UPPER, false, 'FOOBAR'],
-            [['foo', 'bar'], Convert::SPACE, MB_CASE_LOWER, false, 'foo bar'],
-            [['foo', 'bar'], Convert::SPACE, MB_CASE_TITLE, false, 'Foo Bar'],
-            [['foo', 'bar'], Convert::SPACE, MB_CASE_UPPER, false, 'FOO BAR'],
-            [['foo', 'bar'], '§', MB_CASE_LOWER, false, 'foo§bar'],
-            [['foo', 'bar'], '§', MB_CASE_TITLE, false, 'Foo§Bar'],
-            [['foo', 'bar'], '§', MB_CASE_UPPER, false, 'FOO§BAR'],
-            [['foo', 'bar'], Convert::DASH, MB_CASE_LOWER, true, 'foo-bar'],
-            [['foo', 'bar'], Convert::DASH, MB_CASE_TITLE, true, 'foo-Bar'],
-            [['foo', 'bar'], Convert::DASH, MB_CASE_UPPER, true, 'foo-BAR'],
-            [['foo', 'bar'], Convert::UNDERSCORE, MB_CASE_LOWER, true, 'foo_bar'],
-            [['foo', 'bar'], Convert::UNDERSCORE, MB_CASE_TITLE, true, 'foo_Bar'],
-            [['foo', 'bar'], Convert::UNDERSCORE, MB_CASE_UPPER, true, 'foo_BAR'],
-            [['foo', 'bar'], Convert::EMPTY_STRING, MB_CASE_LOWER, true, 'foobar'],
-            [['foo', 'bar'], Convert::EMPTY_STRING, MB_CASE_TITLE, true, 'fooBar'],
-            [['foo', 'bar'], Convert::EMPTY_STRING, MB_CASE_UPPER, true, 'fooBAR'],
-            [['foo', 'bar'], Convert::SPACE, MB_CASE_LOWER, true, 'foo bar'],
-            [['foo', 'bar'], Convert::SPACE, MB_CASE_TITLE, true, 'foo Bar'],
-            [['foo', 'bar'], Convert::SPACE, MB_CASE_UPPER, true, 'foo BAR'],
-            [['foo', 'bar'], '§', MB_CASE_LOWER, true, 'foo§bar'],
-            [['foo', 'bar'], '§', MB_CASE_TITLE, true, 'foo§Bar'],
-            [['foo', 'bar'], '§', MB_CASE_UPPER, true, 'foo§BAR'],
+            'Underscore 1' => [false, UnderscoreSplitter::class, 'hola_mundo'],
+            'Underscore 2' => [false, UnderscoreSplitter::class, 'HELLO_WORLD'],
+            'Underscore 3' => [true, UnderscoreSplitter::class, 'Ñ'],
+            'Underscore 4' => [true, UnderscoreSplitter::class, 'HELLO'],
+            'Uppercase 1'  => [false, UppercaseSplitter::class, ''],
+            'Uppercase 2'  => [false, UppercaseSplitter::class, 'ñ'],
+            'Uppercase 3'  => [false, UppercaseSplitter::class, 'one'],
+            'Uppercase 4'  => [false, UppercaseSplitter::class, 'helloWorld'],
+            'Dash 1'       => [false, DashSplitter::class, 'hello-World'],
+            'Dash 2'       => [false, DashSplitter::class, 'my-name-is-bond'],
+            'Space 1'      => [false, SpaceSplitter::class, 'Hola mundo'],
+            'Space 2'      => [false, SpaceSplitter::class, 'Mi nombre es bond'],
+            'Space 3'      => [false, SpaceSplitter::class, 'Formule courte spéciale été'],
         ];
     }
 
@@ -278,28 +172,35 @@ class ConvertTest extends TestCase
      * @covers       \Jawira\CaseConverter\Convert::toTrain()
      * @covers       \Jawira\CaseConverter\Convert::toUpper()
      *
-     * @param string $converterMethod
+     * @param string $methodName
+     * @param string $className
      *
-     * @throws \ReflectionException
      */
-    public function testConverterMethodCallsGlueString(string $converterMethod)
+    public function testConverterMethodCallsGlueString(string $methodName, string $className)
     {
-        // Disabling constructor without stub methods
-        $mock = $this->getMockBuilder(Convert::class)
-                     ->disableOriginalConstructor()
-                     ->setMethods(['glueString'])
-                     ->getMock();
+        $expected = 'this is a dummy text';
 
-        // Stub called once and returns value
-        $mock->expects($this->once())
-             ->method('glueString');
+        $namingConvention = $this->getMockBuilder($className)
+                                 ->disableOriginalConstructor()
+                                 ->setMethods(['glue'])
+                                 ->getMock();
 
-        // Removing protected for converter method
-        $reflection = new ReflectionObject($mock);
-        $method     = $reflection->getMethod($converterMethod);
-        $method->setAccessible(true);
+        $namingConvention->expects($this->once())
+                         ->method('glue')
+                         ->willReturn($expected);
 
-        $method->invoke($mock);
+        $convertMock = $this->getMockBuilder(Convert::class)
+                            ->disableOriginalConstructor()
+                            ->setMethods(['factory'])
+                            ->getMock();
+
+        $convertMock->expects($this->once())
+                    ->method('factory')
+                    ->willReturn($namingConvention);
+
+        /** @var \Jawira\CaseConverter\Gluer $convertMock */
+        $returned = $convertMock->$methodName();
+        $this->assertSame($expected, $returned);
     }
 
     /**
@@ -308,18 +209,18 @@ class ConvertTest extends TestCase
     public function converterMethodProvider()
     {
         return [
-            'to' . Convert::ADA      => ['to' . Convert::ADA],
-            'to' . Convert::CAMEL    => ['to' . Convert::CAMEL],
-            'to' . Convert::COBOL    => ['to' . Convert::COBOL],
-            'to' . Convert::KEBAB    => ['to' . Convert::KEBAB],
-            'to' . Convert::LOWER    => ['to' . Convert::LOWER],
-            'to' . Convert::MACRO    => ['to' . Convert::MACRO],
-            'to' . Convert::PASCAL   => ['to' . Convert::PASCAL],
-            'to' . Convert::SENTENCE => ['to' . Convert::SENTENCE],
-            'to' . Convert::SNAKE    => ['to' . Convert::SNAKE],
-            'to' . Convert::TITLE    => ['to' . Convert::TITLE],
-            'to' . Convert::TRAIN    => ['to' . Convert::TRAIN],
-            'to' . Convert::UPPER    => ['to' . Convert::UPPER],
+            'toAda'      => ['toAda', AdaCase::class],
+            'toCamel'    => ['toCamel', CamelCase::class],
+            'toCobol'    => ['toCobol', CobolCase::class],
+            'toKebab'    => ['toKebab', KebabCase::class],
+            'toLower'    => ['toLower', LowerCase::class],
+            'toMacro'    => ['toMacro', MacroCase::class],
+            'toPascal'   => ['toPascal', PascalCase::class],
+            'toSentence' => ['toSentence', SentenceCase::class],
+            'toSnake'    => ['toSnake', SnakeCase::class],
+            'toTitle'    => ['toTitle', TitleCase::class],
+            'toTrain'    => ['toTrain', TrainCase::class],
+            'toUpper'    => ['toUpper', UpperCase::class],
         ];
     }
 
@@ -343,144 +244,6 @@ class ConvertTest extends TestCase
         $reflectedClass  = new ReflectionClass(Convert::class);
         $reflectedMethod = $reflectedClass->getMethod('__toString');
         $reflectedMethod->invoke($mock);
-    }
-
-    /**
-     * @covers       \Jawira\CaseConverter\Convert::detectNamingConvention()
-     *
-     * @param string $analyseReturn Expected value returned by analyse() method
-     * @param string $splitMethod   Split method to be called
-     *
-     * @dataProvider detectNamingConventionProvider()
-     *
-     * @throws \ReflectionException
-     */
-    public function testDetectNamingConvention(string $analyseReturn, string $splitMethod)
-    {
-        $inputString = 'deep-space-nine';
-
-        // Disabling constructor and setting stub methods
-        $mock = $this->getMockBuilder(Convert::class)
-                     ->disableOriginalConstructor()
-                     ->setMethods(['analyse', $splitMethod])
-                     ->getMock();
-
-        // set expectations for analyse()
-        $mock->expects($this->once())
-             ->method('analyse')
-             ->with($inputString)
-             ->will($this->returnValue($analyseReturn));
-
-        // set expectations for $splitMethod
-        $mock->expects($this->once())
-             ->method($splitMethod)
-             ->with($inputString);
-
-        // Making public a protected method
-        $reflection = new ReflectionObject($mock);
-        $method     = $reflection->getMethod('detectNamingConvention');
-        $method->setAccessible(true);
-
-        // Testing
-        $output = $method->invoke($mock, $inputString);
-        $this->assertInstanceOf(Convert::class, $output);
-    }
-
-    public function detectNamingConventionProvider()
-    {
-        return [
-            'underscore' => [Convert::STRATEGY_UNDERSCORE, 'splitUnderscoreString'],
-            'dash'       => [Convert::STRATEGY_DASH, 'splitDashString'],
-            'uppercase'  => [Convert::STRATEGY_UPPERCASE, 'splitUppercaseString'],
-            'space'      => [Convert::STRATEGY_SPACE, 'splitSpaceString'],
-        ];
-    }
-
-    /**
-     * Tested methods should call Convert::splitString() method
-     *
-     * @covers       \Jawira\CaseConverter\Convert::splitDashString()
-     * @covers       \Jawira\CaseConverter\Convert::splitUnderscoreString()
-     * @covers       \Jawira\CaseConverter\Convert::splitSpaceString()
-     *
-     * @dataProvider splitStringCallProvider()
-     *
-     * @param string $splitMethod
-     *
-     * @throws \ReflectionException
-     */
-    public function testSplitStringCall(string $splitMethod)
-    {
-        $splitReturnValue = ['this', 'can', 'be', 'anything'];
-
-        // Disabling constructor and setting stub method
-        $mock = $this->getMockBuilder(Convert::class)
-                     ->disableOriginalConstructor()
-                     ->setMethods(['splitString'])
-                     ->getMock();
-
-        // Making public a protected method
-        $reflection = new ReflectionObject($mock);
-        $method     = $reflection->getMethod($splitMethod);
-        $method->setAccessible(true);
-
-        // Expectation
-        $mock->expects($this->once())
-             ->method('splitString')
-             ->will($this->returnValue($splitReturnValue));
-
-        // Testing
-        $output = $method->invoke($mock, $splitMethod);
-        $this->assertSame($splitReturnValue, $output);
-    }
-
-    public function splitStringCallProvider()
-    {
-        return [
-            'splitDashString'       => ['splitDashString'],
-            'splitUnderscoreString' => ['splitUnderscoreString'],
-            'splitSpaceString'      => ['splitSpaceString'],
-        ];
-    }
-
-    /**
-     * @covers       \Jawira\CaseConverter\Convert::splitUppercaseString()
-     * @dataProvider splitUppercaseStringProvider()
-     *
-     * @param string $input
-     * @param string $parameter
-     *
-     * @throws \ReflectionException
-     */
-    public function testSplitUppercaseString(string $input, string $parameter)
-    {
-        // Disabling constructor and setting stub method
-        $mock = $this->getMockBuilder(Convert::class)
-                     ->disableOriginalConstructor()
-                     ->setMethods(['splitUnderscoreString'])
-                     ->getMock();
-
-        // Setting expectation for Stub method
-        $mock->expects($this->once())
-             ->method('splitUnderscoreString')
-             ->with($parameter);
-
-        // Making public a protected method
-        $reflection = new ReflectionObject($mock);
-        $method     = $reflection->getMethod('splitUppercaseString');
-        $method->setAccessible(true);
-
-        // Testing
-        $method->invoke($mock, $input);
-    }
-
-    public function splitUppercaseStringProvider()
-    {
-        return [
-            ['HolaMundo', '_Hola_Mundo'],
-            ['yes', 'yes'],
-            ['airBus', 'air_Bus'],
-        ];
     }
 
     /**
@@ -520,6 +283,120 @@ class ConvertTest extends TestCase
             'small array'  => [['a'], 1],
             'medium array' => [['a', 'a', 'a', 'a',], 4],
             'large array'  => [['a', 'a', 'a', 'a', 'a', 'a', 'a', 'a', 'a', 'a', 'a', 'a',], 12],
+        ];
+    }
+
+    /**
+     * @covers Jawira\CaseConverter\Convert::extractWords
+     *
+     * @throws \ReflectionException
+     */
+    public function testExtractWords()
+    {
+        // Preparing Splitter object
+        $splitterMock = $this->getMockBuilder(DashSplitter::class)
+                             ->disableOriginalConstructor()
+                             ->setMethods(['split'])
+                             ->getMock();
+
+        $splitterMock->expects($this->once())
+                     ->method('split')
+                     ->with()
+                     ->willReturn(['dummy', 'array']);
+
+        // Preparing Convert object
+        $convertMock = $this->getMockBuilder(Convert::class)
+                            ->disableOriginalConstructor()
+                            ->setMethods(['analyse'])
+                            ->getMock();
+
+        $convertMock->expects($this->once())
+                    ->method('analyse')
+                    ->with('dummy-value')
+                    ->will($this->returnValue($splitterMock));
+
+        // Calling protected method
+        $reflection = new ReflectionObject($convertMock);
+        $method     = $reflection->getMethod('extractWords');
+        $method->setAccessible(true);
+        $result = $method->invoke($convertMock, 'dummy-value');
+
+        $this->assertAttributeEquals(['dummy', 'array'], 'words', $convertMock);
+        $this->assertInstanceOf(Convert::class, $result);
+    }
+
+    /**
+     * @covers \Jawira\CaseConverter\Convert::toArray
+     *
+     * @throws \ReflectionException
+     */
+    public function testToArray()
+    {
+        // Preparing Convert mock
+        $mock = $this->getMockBuilder(Convert::class)
+                     ->disableOriginalConstructor()
+                     ->setMethods()
+                     ->getMock();
+
+        // Setting value to protected property
+        $reflectionObject   = new ReflectionObject($mock);
+        $reflectionProperty = $reflectionObject->getProperty('words');
+        $reflectionProperty->setAccessible(true);
+        $reflectionProperty->setValue($mock, ['dummy', 'array']);
+
+        /** @var \Jawira\CaseConverter\Convert $mock */
+        $currentArray = $mock->toArray();
+
+        $this->assertIsArray($currentArray);
+        $this->assertEquals(['dummy', 'array'], $currentArray);
+    }
+
+    /**
+     * @covers       \Jawira\CaseConverter\Convert::factory
+     * @covers       \Jawira\CaseConverter\Gluer::__construct
+     * @dataProvider factoryProvider
+     *
+     * @param string $className
+     *
+     * @throws \ReflectionException
+     */
+    public function testFactory(string $className)
+    {
+        // Preparing Convert mock
+        $mock = $this->getMockBuilder(Convert::class)
+                     ->disableOriginalConstructor()
+                     ->setMethods()
+                     ->getMock();
+
+        // Setting value to protected property
+        $reflectionObject   = new ReflectionObject($mock);
+        $reflectionProperty = $reflectionObject->getProperty('words');
+        $reflectionProperty->setAccessible(true);
+        $reflectionProperty->setValue($mock, ['dummy', 'array']);
+
+        // Calling protected method
+        $reflectionMethod = $reflectionObject->getMethod('factory');
+        $reflectionMethod->setAccessible(true);
+        $result = $reflectionMethod->invoke($mock, $className);
+
+        $this->assertInstanceOf(Gluer::class, $result);
+    }
+
+    public function factoryProvider()
+    {
+        return [
+            [AdaCase::class],
+            [CamelCase::class],
+            [CobolCase::class],
+            [KebabCase::class],
+            [LowerCase::class],
+            [MacroCase::class],
+            [PascalCase::class],
+            [SentenceCase::class],
+            [SnakeCase::class],
+            [TitleCase::class],
+            [TrainCase::class],
+            [UpperCase::class],
         ];
     }
 }